requirements.txt
# CrewAI Framework
crewai==0.102.0
crewai-tools==0.25.8

# LangChain (Ollama-compatible)
langchain==0.3.6
langchain-community==0.3.5
langchain-ollama==0.2.0
langchain-google-genai


# Financial Data
yfinance==0.2.52
alpha-vantage==3.0.0
requests==2.32.3
beautifulsoup4==4.12.3

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Utilities
python-dotenv==1.0.1

# API & Web
httpx==0.27.0
